Added a validation in the delete method of the account service in case the account does not exist, then it returns true

+++ src/main/java/com/kyoshisakuda/budget/account/AccountServiceImpl.java
@@ -42,6 +42,8 @@ public class AccountServiceImpl implements AccountService {
 
     @Override
     public boolean deleteAccount(int id) {
+        if (!repository.findById(id).isPresent())
+            return true;
         repository.deleteById(id);
         return !repository.findById(id).isPresent();
     }
